+++ ci/1-build/macos/build.sh
@@ -6,7 +6,8 @@ repo_root=$(git rev-parse --show-toplevel)
 
 pushd "$repo_root/packages/nesd" >/dev/null
 
-flutter build macos --release --flavor "$FLAVOR"
+flutter build macos --release --flavor "$FLAVOR" \
+  --dart-define=NESD_BUILD_ID="${NESD_BUILD_ID:-}"
 
 popd >/dev/null